refactor: 리뷰 반영

+++ src/main/java/techcourse/myblog/web/CommentController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
+import techcourse.myblog.domain.Comment;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.dto.CommentSaveRequestDto;
 import techcourse.myblog.service.CommentService;
@@ -17,10 +18,13 @@ import static techcourse.myblog.util.SessionKeys.USER;
 
 @Slf4j
 @Controller
-@RequiredArgsConstructor
 public class CommentController {
 
     private final CommentService commentService;
+
+    public CommentController(final CommentService commentService) {
+        this.commentService=commentService;
+    }
 
     @PostMapping("/comment/writing")
     public String saveComment(CommentSaveRequestDto commentSaveRequestDto, HttpSession httpSession) {
@@ -30,18 +34,26 @@ public class CommentController {
     }
 
     @PutMapping("/comment/edit/{id}")
-    public String editComment(@PathVariable Long id, String editedContents) {
-        commentService.update(id, editedContents);
-
+    public String editComment(@PathVariable Long id, String editedContents, HttpSession httpSession) {
         Long articleId = commentService.findArticleIdById(id);
-
+        if (checkUserIsCommentAuthor(id, httpSession)) {
+            commentService.update(id, editedContents);
+        }
         return "redirect:/articles/" + articleId;
     }
 
     @DeleteMapping("/comment/edit/{id}")
-    public String deleteComment(@PathVariable Long id) {
+    public String deleteComment(@PathVariable Long id, HttpSession httpSession) {
         Long articleId = commentService.findArticleIdById(id);
-        commentService.deleteById(id);
+        if (checkUserIsCommentAuthor(id, httpSession)) {
+            commentService.deleteById(id);
+        }
         return "redirect:/articles/" + articleId;
+    }
+
+    private boolean checkUserIsCommentAuthor(Long id, HttpSession httpSession) {
+        Comment comment = commentService.findById(id);
+        User user = (User) httpSession.getAttribute(USER);
+        return comment.isAuthor(user);
     }
 }
